Split chroma actions to Memories.py
